+++ src/main/java/syleelsw/anyonesolveit/api/study/dto/SolvedacItem.java
@@ -2,12 +2,13 @@ package syleelsw.anyonesolveit.api.study.dto;
 
 import lombok.Getter;
 import lombok.ToString;
+import syleelsw.anyonesolveit.service.study.dto.DisplayName;
 
 import java.util.List;
 
 @Getter @ToString
 public class SolvedacItem {
-    private Integer problemId;
+    private Long problemId;
     private String titleKo;
     private Integer level;
     private List<SolvedTag> tags;
